pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "django-bugz"
version = "0.1.0"
description = ""
authors = ["Alexandre Macabies <web+oss@zopieux.com>"]
packages = [{ include = "bugz"}]
repository = "https://github.com/zopieux/django-bugz"

[tool.poetry.dependencies]
python = ">=3.8"
django = ">=3,<4"
markdown = ">=3,<4"
bleach = ">=3,<4"
rules = ">=2,<3"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"
